test : 테스트를 위한 TestCar 객체 분리 및 테스트 케이스 추가
